Make SerializedDatasetRefContainerV1 public
